requirements.txt
pandas~=1.4.1
google-api-python-client~=2.36.0
oauth2client~=4.1.3